Can now add an action for each inputted character type
Currently not many error checks - need to add checks if user changes a character under action pairs to one that isn't stored. Or if there is a double character entry. Or if a character is added to action map but not to mutation rules - need to add checks to make edits in both action map and mutation rules cross compatible without causing errors.