Rearranged LStringReader to Only Instantiate on Grow Commands
LStringReader now checks to make sure the last read character of the string is not a grow character. If it is, then it does not instantiate at a random position. If it is not, then it does instantiate at a random position. Instantiated growth also currently uses angles defined by characters growth angle mapping